[Fix-17420] fix oauth2 login failed
