+++ dolphinscheduler-api-test/dolphinscheduler-api-test-case/pom.xml
@@ -28,10 +28,6 @@
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <properties>
-        <selenium.version>4.21.0</selenium.version>
-    </properties>
-
     <artifactId>dolphinscheduler-api-test-case</artifactId>
 
     <dependencies>
@@ -49,17 +45,6 @@
             <groupId>org.apache.dolphinscheduler</groupId>
             <artifactId>dolphinscheduler-api</artifactId>
             <version>dev-SNAPSHOT</version>
-        </dependency>
-        <dependency>
-            <groupId>org.seleniumhq.selenium</groupId>
-            <artifactId>selenium-chrome-driver</artifactId>
-            <version>${selenium.version}</version>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.seleniumhq.selenium</groupId>
-            <artifactId>selenium-support</artifactId>
-            <version>${selenium.version}</version>
         </dependency>
     </dependencies>
 </project>
