+++ dolphinscheduler-e2e/dolphinscheduler-e2e-case/src/test/resources/docker/oauth-login/application.yaml
